MODIFY contact with backend for duplicate elements

+++ admin/liste-employes.php
@@ -1,6 +1,5 @@
 <?php
 require_once __DIR__ . "/../lib/config.php";
-
 require_once __DIR__ . "/../lib/session.php";
 //only admin has permission to visit this page
 adminOnly();
@@ -37,7 +36,6 @@ $totalPages = ceil($totalEmployees / _ADMIN_ITEM_PER_PAGE_);
       <div class="connection-table">
         <table>
           <tr>
-            <!-- <th>#</th> -->
             <th>Nom</th>
             <th>Prénom</th>
             <th>Email</th>
@@ -46,21 +44,18 @@ $totalPages = ceil($totalEmployees / _ADMIN_ITEM_PER_PAGE_);
           </tr>
           <tbody>
             <?php foreach ($employees as $employee) { ?>
-            <tr>
-              <!-- <th scope="row"><?= $employee["id"]; ?></th> -->
-              <td><?= $employee["lastname"]; ?></td>
-              <td><?= $employee["name"]; ?></td>
-              <td><?= $employee["email"]; ?></td>
-              <td><?= $employee["role"]; ?></td>
+              <tr>
+                <td><?= $employee["lastname"]; ?></td>
+                <td><?= $employee["name"]; ?></td>
+                <td><?= $employee["email"]; ?></td>
+                <td><?= $employee["role"]; ?></td>
 
-              <?php if ($employee["role"] === 'employee') { ?>
-              <td><a href="modifier-employe.php?id=<?= $employee['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
-                | <a href="delete-employe.php?id=<?= $employee['id'] ?>"
-                  onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i
-                    class="fa-solid fa-trash-can"></i></a>
-                <?php } ?>
-              </td>
-            </tr>
+                <?php if ($employee["role"] === 'employee') { ?>
+                  <td><a href="modifier-employe.php?id=<?= $employee['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
+                    | <a href="delete-employe.php?id=<?= $employee['id'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i class="fa-solid fa-trash-can"></i></a>
+                  <?php } ?>
+                  </td>
+              </tr>
             <?php } ?>
           </tbody>
         </table>
@@ -69,15 +64,15 @@ $totalPages = ceil($totalEmployees / _ADMIN_ITEM_PER_PAGE_);
       <nav aria-label="Page navigation employes">
         <ul class="pagination">
           <?php if ($totalPages > 1) { ?>
-          <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
-          <li class="page-item">
-            <a class="page-link <?php if ($i == $page) {
+            <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
+              <li class="page-item">
+                <a class="page-link <?php if ($i == $page) {
                                       echo " active";
                                     } ?>" href="?page=<?php echo $i; ?>">
-              <?php echo $i; ?>
-            </a>
-          </li>
-          <?php } ?>
+                  <?php echo $i; ?>
+                </a>
+              </li>
+            <?php } ?>
           <?php } ?>
         </ul>
       </nav>
